pyproject.toml
[project]
name = "internagent-1022"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aider-chat==0.86.1",
    "aider-install==0.2.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anthropic==0.69.0",
    "anyio==4.10.0",
    "argcomplete==3.6.2",
    "attrs==25.3.0",
    "backoff==2.2.1",
    "beautifulsoup4==4.13.4",
    "cachetools==5.5.2",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "ConfigArgParse==1.7.1",
    "contourpy==1.3.3",
    "cycler==0.12.1",
    "diff-match-patch==20241021",
    "diskcache==5.6.3",
    "distro==1.9.0",
    "docstring_parser==0.17.0",
    "dotenv==0.9.9",
    "easydict==1.13",
    "filelock==3.18.0",
    "flake8==7.3.0",
    "fonttools==4.60.1",
    "frozenlist==1.7.0",
    "fsspec==2025.7.0",
    "gitdb==4.0.12",
    "GitPython==3.1.45",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.25.1",
    "google-api-python-client==2.178.0",
    "google-auth==2.40.3",
    "google-auth-httplib2==0.2.0",
    "google-generativeai==0.8.5",
    "googleapis-common-protos==1.70.0",
    "graphviz==0.21",
    "grep-ast==0.9.0",
    "grpcio==1.74.0",
    "grpcio-status==1.71.2",
    "h11==0.16.0",
    "hf-xet==1.1.7",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "huggingface-hub==0.34.3",
    "idna==3.10",
    "importlib_metadata==7.2.1",
    "importlib_resources==6.5.2",
    "Jinja2==3.1.6",
    "jiter==0.10.0",
    "json5==0.12.0",
    "json_repair==0.51.0",
    "jsonschema==4.25.0",
    "jsonschema-specifications==2025.4.1",
    "kiwisolver==1.4.9",
    "litellm==1.75.0",
    "lxml==6.0.2",
    "markdown-it-py==3.0.0",
    "MarkupSafe==3.0.2",
    "matplotlib==3.10.6",
    "mccabe==0.7.0",
    "mdurl==0.1.2",
    "mixpanel==4.10.1",
    "mpmath==1.3.0",
    "mslex==1.3.0",
    "multidict==6.6.3",
    "networkx==3.4.2",
    "numpy==1.26.4",
    "nvidia-cublas-cu12==12.8.4.1",
    "nvidia-cuda-cupti-cu12==12.8.90",
    "nvidia-cuda-nvrtc-cu12==12.8.93",
    "nvidia-cuda-runtime-cu12==12.8.90",
    "nvidia-cudnn-cu12==9.10.2.21",
    "nvidia-cufft-cu12==11.3.3.83",
    "nvidia-cufile-cu12==1.13.1.3",
    "nvidia-curand-cu12==10.3.9.90",
    "nvidia-cusolver-cu12==11.7.3.90",
    "nvidia-cusparse-cu12==12.5.8.93",
    "nvidia-cusparselt-cu12==0.7.1",
    "nvidia-nccl-cu12==2.27.3",
    "nvidia-nvjitlink-cu12==12.8.93",
    "nvidia-nvtx-cu12==12.8.90",
    "openai==1.99.1",
    "oslex==0.1.3",
    "packaging==25.0",
    "pathspec==0.12.1",
    "pdfminer.six==20231228",
    "pdfplumber==0.11.5",
    "pexpect==4.9.0",
    "pillow==11.3.0",
    "pipx==1.7.1",
    "platformdirs==4.4.0",
    "posthog==6.4.1",
    "prompt_toolkit==3.0.51",
    "propcache==0.3.2",
    "proto-plus==1.26.1",
    "protobuf==5.29.5",
    "psutil==7.0.0",
    "ptyprocess==0.7.0",
    "pyasn1==0.6.1",
    "pyasn1_modules==0.4.2",
    "pycodestyle==2.14.0",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic_core==2.33.2",
    "pydub==0.25.1",
    "pyflakes==3.4.0",
    "Pygments==2.19.2",
    "pypandoc==1.15",
    "pyparsing==3.2.3",
    "pypdf==6.1.0",
    "pypdfium2==4.30.0",
    "pyperclip==1.9.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "PyYAML==6.0.2",
    "referencing==0.36.2",
    "regex==2025.7.34",
    "requests==2.32.4",
    "rich==14.1.0",
    "rpds-py==0.27.0",
    "rsa==4.9.1",
    "scipy==1.15.3",
    "shtab==1.7.2",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "socksio==1.0.0",
    "sounddevice==0.5.2",
    "soundfile==0.13.1",
    "soupsieve==2.7",
    "sympy==1.14.0",
    "tiktoken==0.10.0",
    "tokenizers==0.21.4",
    "torch==2.8.0",
    "tqdm==4.67.1",
    "tree-sitter==0.24.0",
    "tree-sitter-c-sharp==0.23.1",
    "tree-sitter-embedded-template==0.23.2",
    "tree-sitter-language-pack==0.9.0",
    "tree-sitter-yaml==0.7.1",
    "triton==3.4.0",
    "typing-inspection==0.4.1",
    "typing_extensions==4.14.1",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
    "userpath==1.9.2",
    "watchfiles==1.1.0",
    "wcwidth==0.2.13",
    "websockets==15.0.1",
    "yarl==1.20.1",
    "zipp==3.23.0",
]